All Bug Changes
The first fix was that if n  < 1 then the program would return false, but "false" doesn't work in c, any number other than 0 is false so I chnaged it to return0; .   Then I noticed that this program was using numberOfDigits, which isn't declared, instead of numDigits.  After that this program had two different spellings of modulus so i edited 'modulous' to modulus so it was the same variable.  Lastly, I noticed there were 3 tests failing, so i checked the equations and noticed that instead of using the log of square, this program used n which is the original number.